Fix scrollbar and width issue on drawers in IE9
IE9 has a bug that subtracts the width of the scrollbars from the
defined width on fixed position elements with box-sizing: border-box.
The workaround is to use content-box instead (on the drawer only). This
means that we can't apply a border or padding to the drawer element
without messing up the off-canvas layout. Instead, padding should be
applied to a nested div.